Seperate the TFM RLE Reader and primitives into a different header file
